mobile-emulation/requirements.txt
Appium-Python-Client==4.3.0
selenium==4.27.1
openai==1.82.0
httpx==0.28.1
python-dotenv==1.0.1
rich==13.9.4
click==8.1.7
